Moved sparce_matrix from data_structures to data_structures/matrix/cpp
